[refactor] Configuration 네이밍 통일, 주석 제거
