+++ src/main/java/com/prgrms/rg/config/WebConfiguration.java
@@ -1,0 +1,17 @@
+package com.prgrms.rg.config;
+
+import com.prgrms.rg.web.LoggingInterceptor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfiguration implements WebMvcConfigurer {
+
+  @Override
+  public void addInterceptors(InterceptorRegistry registry) {
+    registry.addInterceptor(new LoggingInterceptor())
+        .order(1)
+        .addPathPatterns("/**");
+  }
+}
